Arreglo excepciones, elimino clases inncesesarias, refactor en general

+++ src/main/java/com/utn/TPFUDEE/Services/AddressService.java
@@ -1,9 +1,5 @@
 package com.utn.TPFUDEE.Services;
 
-import com.utn.TPFUDEE.Exceptions.Exist.AddressExistException;
-import com.utn.TPFUDEE.Exceptions.Exist.UserExistException;
-import com.utn.TPFUDEE.Exceptions.NoContent.AddressNoContentException;
-import com.utn.TPFUDEE.Exceptions.NotFound.AddressNotFoundException;
 import com.utn.TPFUDEE.Models.Address;
 import com.utn.TPFUDEE.Repositories.AddressRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,10 +7,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.server.ResponseStatusException;
 
-import java.util.ArrayList;
-import java.util.List;
 
 @Service
 public class AddressService {
@@ -26,35 +20,31 @@ public class AddressService {
         this.addressRepository = addressRepository;
     }
 
-    public Page<Address> getAll(Pageable pageable) throws AddressNoContentException{
+    public Page<Address> getAll(Pageable pageable){
         Page<Address> addressList = addressRepository.findAll(pageable);
         if(addressList.isEmpty()){
-            throw new AddressNoContentException();
+            throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Empty List");
         }
         return addressList;
     }
 
-    public Address add(Address address) throws AddressExistException {
-        boolean flag = false;
-        for(Address var : addressRepository.findAll()){
-            if(address.getStreet().equals(var.getStreet()) && address.getNumber().equals(var.getNumber())){
-                flag = true;
-            }
-        }
-        if(flag){
-            throw new AddressExistException();
+    public Address add(Address address){
+        if((addressRepository.findByStreetAndNumber(address.getStreet(), address.getNumber()))!=null){
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Address Already Exist");
         }else{
             return addressRepository.save(address);
         }
     }
 
-    public Address getById(Integer id) throws AddressNotFoundException{
-        return addressRepository.findById(id).orElseThrow(() -> new AddressNotFoundException());
+    public Address getById(Integer id){
+        return addressRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Address Not Found"));
     }
 
-    public void deleteById(Integer id) throws AddressNotFoundException{
-        this.getById(id);
-        addressRepository.deleteById(id);
+    public void deleteById(Integer id){
+        if((this.getById(id))!=null)
+            addressRepository.deleteById(id);
+        else
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Address Not Found For delete");
     }
